UFBA/LLMs/pyproject.toml
[tool.poetry]
name = "LLMs"
version = "0.1.0"
description = ""
authors = ["Silas Henrique <silash35@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.taskipy.tasks]
main = "python LLMs/main.py"
chat = "python LLMs/chat.py"
filter = "python LLMs/filter.py"


[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.2"
langchain-huggingface = "^0.0.3"
langchain-community = "^0.2.3"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.8"
pypdf = "^4.2.0"
youtube-transcript-api = "^0.6.2"
langchain-text-splitters = "^0.2.1"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
taskipy = "^1.12.2"
watchdog = "^4.0.1"
huggingface-cli = "^0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
